Reorder files.
